Set unified_mode on resources for Chef 17 support
Require Chef 15.3+ for unified_mode support
