--- .rubocop.yml
@@ -1,16 +1,0 @@
-# rubocop todo
-
-Lint/SuppressedException:
-  Exclude:
-    - 'libraries/torrent.rb'
-
-Lint/Loop:
-  Exclude:
-    - 'providers/torrent_file.rb'
-
-Metrics/BlockNesting:
-  Max: 4
-
-Style/RegexpLiteral:
-  Exclude:
-    - 'providers/torrent_file.rb'
